requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.2
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
packaging==21.0
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
py-solc-x==1.1.0
pycryptodome==3.10.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
requests==2.26.0
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
toml==0.10.2
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
web3==5.23.1
websockets==9.1
yarl==1.6.3
